created separate modal page
moved modals to separate page so it's less cluttered
