+++ frontend/src/todos/completedTodos.js
@@ -14,17 +14,12 @@ import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 import Paper from '@mui/material/Paper';
 import Checkbox from '@mui/material/Checkbox';
-import IconButton from '@mui/material/IconButton';
 import Button from '@mui/material/Button';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import Switch from '@mui/material/Switch';
 import { visuallyHidden } from '@mui/utils';
 import { useHistory } from "react-router-dom";
-import EditIcon from '@mui/icons-material/Edit';
-import Stack from '@mui/material/Stack';
-import Modal from '@mui/material/Modal';
-import RemoveDoneIcon from '@mui/icons-material/RemoveDone';
-import DoneAllIcon from '@mui/icons-material/DoneAll';
+import { RemoveFromCompleteModal } from "./completeModals";
 //TO REDIRECT TO CONFIRMATION PAGE OR OTHER PAGES 
 // it works sometimes... its a little odd 
 //  <Redirect to = "/confirmation"/>
@@ -39,17 +34,6 @@ function createData(task, duedate, actual) {
 }
 
 
-const style = {
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    width: 400,
-    bgcolor: 'background.paper',
-    border: '2px solid #000',
-    boxShadow: 24,
-    p: 4,
-  };
 const rows = [
     createData("Start 307 Project", "10/25/2021", "10/17/2021"),
     createData("Get Groceries", "10/30/2021", "10/30/2021"),
@@ -167,9 +151,6 @@ const EnhancedTableToolbar = (props) => {
 
   
 
-    const [open, setOpen] = React.useState(false);
-    const handleOpen = () => setOpen(true);
-    const handleClose = () => setOpen(false);
     const history = useHistory();
 
     const handleRoute = () => {
@@ -177,10 +158,6 @@ const EnhancedTableToolbar = (props) => {
     }
 
     const { numSelected } = props;
-    console.log(props)
-    const handleRouteT = () =>{ 
-        history.push("/todos");
-    }
 
     return (
         <Toolbar
@@ -221,31 +198,7 @@ const EnhancedTableToolbar = (props) => {
             )}
 
             {numSelected > 0 ? (
-                <div>
-                <IconButton >
-                    <EditIcon onClick={handleOpen}/>     
-                </IconButton>
-                <Modal
-                    open={open}
-                    onClose={handleClose}
-                    aria-labelledby="modal-modal-title"
-                    aria-describedby="modal-modal-description"
-                >
-                <Box sx={style}>
-                  <Typography id="modal-modal-title" variant="h5" component="h2">
-                    Add this task back to your todo list?
-                    <Stack direction="column" spacing={4}>
-                        <Button variant="contained" style={{ height: '45px', width: '310px', top: 10, left: 45 }} startIcon={<DoneAllIcon />}  onClick={handleRouteT}>
-                            Yes, add to my todo list
-                        </Button>
-                        <Button variant="contained" style={{ height: '45px', width: '310px', top: 10, left: 45 }} startIcon={<RemoveDoneIcon />} onClick={handleClose}>
-                            No, leave as complete
-                        </Button>
-                    </Stack>
-                  </Typography>
-                </Box>
-              </Modal>
-            </div>
+                <RemoveFromCompleteModal/>
             ) : (
                 <div>
